requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fuzzywuzzy==0.18.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
Levenshtein==0.26.1
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.64.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
portalocker==2.10.1
protobuf==5.29.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pylance==0.23.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-Levenshtein==0.26.1
python-socketio==5.12.1
pytz==2025.1
pywin32==308
qdrant-client==1.13.2
RapidFuzz==3.12.1
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
slack_bolt==1.22.0
slack_sdk==3.34.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uuid==1.30
Werkzeug==3.1.3
wsproto==1.2.0
